Fix #6.
Transformation failed when source (input) and target (output) paths were the same, because it tried to open the output file for writing while the input file was still open, leading to an 'The process cannot access the file because it is being used by another process' exception. Now closes the input file before writing the output.